refactor(NotificationReceiver): 이름 변경
- 이전: NotificationReceiver
- 이후: ScreeningNotificationReceiver
