+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             android:parentActivityName=".feature.reservation.ReservationActivity"
             android:exported="false" />
 
-        <receiver android:name=".feature.notification.NotificationReceiver"
+        <receiver android:name=".feature.notification.ScreeningNotificationReceiver"
             android:enabled="true"
             android:exported="false">
             <intent-filter>
