+++ app/src/main/java/woowacourse/movie/feature/notification/ScreeningNotificationReceiver.kt
@@ -3,13 +3,13 @@ package woowacourse.movie.feature.notification
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import woowacourse.movie.db.ticket.Ticket.Companion.DEFAULT_TICKET_ID
+import woowacourse.movie.db.ticket.TicketEntity.Companion.DEFAULT_TICKET_ID
 import woowacourse.movie.feature.finished.ReservationFinishedActivity
 import woowacourse.movie.feature.history.ReservationHistoryFragment.Companion.TICKET_ID
-import woowacourse.movie.feature.notification.ScreeningAlarm.Companion.NOTIFICATION_TEXT
-import woowacourse.movie.feature.notification.ScreeningAlarm.Companion.NOTIFICATION_TITLE
+import woowacourse.movie.feature.notification.ScreeningAlarm.Companion.SCREENING_NOTIFICATION_TEXT
+import woowacourse.movie.feature.notification.ScreeningAlarm.Companion.SCREENING_NOTIFICATION_TITLE
 
-class NotificationReceiver : BroadcastReceiver() {
+class ScreeningNotificationReceiver : BroadcastReceiver() {
     override fun onReceive(
         context: Context,
         intent: Intent,
@@ -17,8 +17,8 @@ class NotificationReceiver : BroadcastReceiver() {
         val notification = ScreeningNotification(context)
 
         val ticketId = intent.getLongExtra(TICKET_ID, DEFAULT_TICKET_ID)
-        val title = intent.getStringExtra(NOTIFICATION_TITLE)
-        val description = intent.getStringExtra(NOTIFICATION_TEXT)
+        val title = intent.getStringExtra(SCREENING_NOTIFICATION_TITLE)
+        val description = intent.getStringExtra(SCREENING_NOTIFICATION_TEXT)
 
         if (title != null && description != null) {
             val newIntent = Intent(context, ReservationFinishedActivity::class.java).putExtra(TICKET_ID, ticketId)
